requirements.txt
Flask>=2.0
Flask-SocketIO>=5.0
eventlet>=0.33.0
python-dotenv>=0.19.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.0.0
Werkzeug>=2.2.0